Refactored af, working and estable version
