+++ .vscode/settings.json
@@ -1,0 +1,72 @@
+{
+    "workbench.colorCustomizations": {
+        "activityBar.activeBackground": "#fa1b49",
+        "activityBar.activeBorder": "#155e02",
+        "activityBar.background": "#fa1b49",
+        "activityBar.foreground": "#e7e7e7",
+        "activityBar.inactiveForeground": "#e7e7e799",
+        "activityBarBadge.background": "#155e02",
+        "activityBarBadge.foreground": "#e7e7e7",
+        "statusBar.background": "#dd0531",
+        "statusBar.foreground": "#e7e7e7",
+        "statusBarItem.hoverBackground": "#fa1b49",
+        "titleBar.activeBackground": "#dd0531",
+        "titleBar.activeForeground": "#e7e7e7",
+        "titleBar.inactiveBackground": "#dd053199",
+        "titleBar.inactiveForeground": "#e7e7e799"
+    },
+    "peacock.color": "#dd0531",
+    "files.associations": {
+        "algorithm": "c",
+        "atomic": "c",
+        "cctype": "c",
+        "cmath": "c",
+        "compare": "c",
+        "concepts": "c",
+        "cstddef": "c",
+        "cstdint": "c",
+        "cstdio": "c",
+        "cstdlib": "c",
+        "cstring": "c",
+        "ctime": "c",
+        "cwchar": "c",
+        "deque": "c",
+        "exception": "c",
+        "initializer_list": "c",
+        "ios": "c",
+        "iosfwd": "c",
+        "iostream": "c",
+        "istream": "c",
+        "iterator": "c",
+        "limits": "c",
+        "locale": "c",
+        "memory": "c",
+        "new": "c",
+        "ostream": "c",
+        "queue": "c",
+        "regex": "c",
+        "stdexcept": "c",
+        "streambuf": "c",
+        "string": "c",
+        "system_error": "c",
+        "tuple": "c",
+        "type_traits": "c",
+        "typeinfo": "c",
+        "utility": "c",
+        "vector": "c",
+        "xfacet": "c",
+        "xiosbase": "c",
+        "xlocale": "c",
+        "xlocbuf": "c",
+        "xlocinfo": "c",
+        "xlocmes": "c",
+        "xlocmon": "c",
+        "xlocnum": "c",
+        "xloctime": "c",
+        "xmemory": "c",
+        "xstddef": "c",
+        "xstring": "c",
+        "xtr1common": "c",
+        "xutility": "c"
+    }
+}
